* development: (28 commits)
  feat: covenants implementation (#3656)
  ci: add tor script to binaries bundle (#3689)
  chore: remove testnet reset todo in wallet (#3685)
  feat: dibbler new genesis block with faucet utxos (#3688)
  ci: fix clippy warning on generated proto module (#3690)
  test: fix metadata signature cucumber (#3687)
  refactor!: clean up #testnet reset TODOs (#3682)
  feat(comms)!: add signature to peer identity to allow third party identity updates (#3629)
  chore: remove moving lock.mdb (#3674)
  chore: merge weatherwax
  feat!: provide a compact form of TransactionInput (#3460)
  fix: allow 0-conf in blockchain db (#3680)
  v0.22.1.1
  v0.22.1
  ci: add build step (#3678)
  test: fix cucumber WalletQuery.feature (#3677)
  test: fix `wait for` step (#3673)
  fix: edge cases causing bans during header/block sync (#3661)
  perf(comms)!: optimise connection establishment (#3658)
  fix: end stale outbound queue immediately on disconnect, retry outbound messages (#3664)
  ...

+++ base_layer/core/src/test_helpers/blockchain.rs
@@ -150,6 +150,7 @@ pub fn create_test_db() -> TempDatabase {
 pub struct TempDatabase {
     path: PathBuf,
     db: Option<LMDBDatabase>,
+    delete_on_drop: bool,
 }
 
 impl TempDatabase {
@@ -159,7 +160,21 @@ impl TempDatabase {
         Self {
             db: Some(create_lmdb_database(&temp_path, LMDBConfig::default()).unwrap()),
             path: temp_path,
+            delete_on_drop: true,
         }
+    }
+
+    pub fn from_path<P: AsRef<Path>>(temp_path: P) -> Self {
+        Self {
+            db: Some(create_lmdb_database(&temp_path, LMDBConfig::default()).unwrap()),
+            path: temp_path.as_ref().to_path_buf(),
+            delete_on_drop: true,
+        }
+    }
+
+    pub fn disable_delete_on_drop(&mut self) -> &mut Self {
+        self.delete_on_drop = false;
+        self
     }
 }
 
@@ -181,7 +196,7 @@ impl Drop for TempDatabase {
     fn drop(&mut self) {
         // force a drop on the LMDB db
         self.db = None;
-        if Path::new(&self.path).exists() {
+        if self.delete_on_drop && Path::new(&self.path).exists() {
             fs::remove_dir_all(&self.path).expect("Could not delete temporary file");
         }
     }
